__init__.py
"""
ML Training Optimizer Environment.

An OpenEnv environment where AI agents learn to optimize ML model training
by tuning hyperparameters, selecting optimizers, and managing training strategies
on real PyTorch models.
"""

from openenv.core.env_server.mcp_types import CallToolAction, ListToolsAction

try:
    from .client import MLTrainerEnv
except ImportError:
    from client import MLTrainerEnv

__all__ = ["MLTrainerEnv", "CallToolAction", "ListToolsAction"]
