pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-ml-trainer-env"
version = "0.1.0"
description = "ML Training Optimizer Environment for OpenEnv - train real PyTorch models with agent-driven hyperparameter optimization"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime
    "openenv-core[core]>=0.2.2",
    # ML training dependencies
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    # Server dependencies
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    # Inference script dependencies
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "ml_trainer_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["ml_trainer_env", "ml_trainer_env.server"]
package-dir = { "ml_trainer_env" = ".", "ml_trainer_env.server" = "server" }